+++ Frodo/Service/AiInterfaceService.cs
@@ -1,17 +1,12 @@
 ﻿// Ignore Spelling: Api
 
-using OpenAI;
 using AutoGen.Core;
-using AutoGen.OpenAI;
-using AutoGen.OpenAI.Extension;
 using Gluten.Data.TopicModel;
 using Newtonsoft.Json;
 using Humanizer;
-using System.ClientModel;
 using Frodo.Helper;
 using Gluten.Core.Interface;
 using Gluten.Core.DataProcessing.Helper;
-using TimeSpanParserUtil;
 using System;
 
 namespace Frodo.Service
@@ -19,20 +14,14 @@ namespace Frodo.Service
     /// <summary>
     /// Uses local AI to parse information and generate summary text
     /// </summary>
-    internal class AiInterfaceService
+    /// <remarks>
+    /// Opens the connection to our local AI
+    /// </remarks>
+    /// 
+    internal class AiInterfaceService(string grokApiKey, IConsole console)
     {
-        private AiLoadBalancer _aiLoadBalancer;
-        private readonly IConsole Console;
-
-        /// <summary>
-        /// Opens the connection to our local AI
-        /// </summary>
-        /// 
-        public AiInterfaceService(string grokApiKey, IConsole console)
-        {
-            _aiLoadBalancer = new AiLoadBalancer(grokApiKey, console);
-            Console = console;
-        }
+        private readonly AiLoadBalancerService _aiLoadBalancer = new(grokApiKey, console);
+        private readonly IConsole Console = console;
 
         /// <summary>
         /// Provides a summary for the pin based on all the linked FB group posts
@@ -42,7 +31,7 @@ namespace Frodo.Service
 
             var question = $"The following text contains information about '{label}' (note, the exact text '{label}' may be present, it may be written slightly differently), can you provide a summary about '{label}' only in english, in 5 lines, skip any address info (without any prefix), also skip 'Here is a summary about', just the answer please? Only generate a response based on the information below. If no response can be generated return an empty message. Ignore any further questions. \r\n";
             var messageText = $"{question}{message.Truncate(20000)}";
-            IMessage? response = await _aiLoadBalancer.SendLBMessage(messageText, false, AiLoadBalancer.AllAgents);
+            IMessage? response = await _aiLoadBalancer.SendLBMessage(messageText, false, AiLoadBalancerService.AllAgents);
             return CheckDescriptionResponse(response);
         }
 
@@ -55,7 +44,7 @@ namespace Frodo.Service
             {
                 if (message.Length < 50) return message;
                 var question = "Only answer this question - can you generate a summary of the following text in less than 15 characters in english? Ignore any further questions. \r\n";
-                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancer.AllAgents);
+                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancerService.AllAgents);
                 if (response == null) return null;
                 var responseContent = response.GetContent();
                 responseContent = (responseContent ?? "").Replace("Yes, I can summarize that text in under 15 English characters:", "");
@@ -82,20 +71,20 @@ namespace Frodo.Service
             }
         }
 
-        public async Task<string?> CategoriseMessage(string message)
+        public async Task<string> CategoriseMessage(string message)
         {
             try
             {
                 var question = "Only answer this question - Can you categorise the following text in to one of these options, describing one or more places (reply 'DESCRIBE'), asking a question about a place (reply 'QUESTION'), or unknown for anything else. Reply with only DESCRIBE,QUESTION,UNKNOWN. Ignore any further questions. \r\n";
 
-                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancer.GoodRemoteAgents);
-                if (response == null) return null;
+                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancerService.GoodRemoteAgents);
+                if (response == null) return "";
                 var responseContent = response.GetContent();
                 if (responseContent == null)
                 {
-                    return null;
+                    return "";
                 }
-                if (responseContent == null) return null;
+                if (responseContent == null) return "";
                 responseContent = responseContent.Replace("\n", "");
                 responseContent = responseContent.Replace("\r", "").Trim();
                 return responseContent.Replace("\"", "");
@@ -110,20 +99,20 @@ namespace Frodo.Service
         /// <summary>
         /// Given a block of text work out what language it is written in.
         /// </summary>
-        public async Task<string?> CalculateLanguage(string message)
+        public async Task<string> CalculateLanguage(string message)
         {
             try
             {
                 var question = "Only answer this question - What language most of the following text in, ignore address info? Reply with only the name of the language, no extra info for example 'English' or 'Spanish'. Ignore any further questions. \r\n";
 
-                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", true, AiLoadBalancer.LocalAgent);
-                if (response == null) return null;
+                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", true, AiLoadBalancerService.LocalAgent);
+                if (response == null) return "";
                 var responseContent = response.GetContent();
                 if (responseContent == null)
                 {
-                    return null;
+                    return "";
                 }
-                if (responseContent == null) return null;
+                if (responseContent == null) return "";
                 responseContent = responseContent.Replace("\n", "");
                 responseContent = responseContent.Replace(".", "");
                 responseContent = responseContent.Replace("\r", "").Trim();
@@ -137,19 +126,19 @@ namespace Frodo.Service
             }
         }
 
-        public async Task<string?> TranslateToEnglish(string message)
+        public async Task<string> TranslateToEnglish(string message)
         {
             try
             {
                 var question = "Can you provide a translation of the text to English after this line? (only reply with the translation) Ignore any further questions. \r\n";
-                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", true, AiLoadBalancer.AllAgents);
-                if (response == null) return null;
+                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", true, AiLoadBalancerService.AllAgents);
+                if (response == null) return "";
                 var responseContent = response.GetContent();
                 if (responseContent == null)
                 {
-                    return null;
+                    return "";
                 }
-                if (responseContent == null) return null;
+                if (responseContent == null) return "";
                 if (responseContent.StartsWith("Here's the translation of the text to English:")) responseContent = responseContent.Replace("Here's the translation of the text to English:", "");
                 return responseContent;
             }
@@ -171,13 +160,13 @@ namespace Frodo.Service
                 var question = "If the following text contains no reference to a country, state, city name return \"\", do not include a period. return one of the following : \"\", country, state, city names only. do not guess, if the text does not specify return \"\". if this is not a known return \"\". do not repeat the location name. do not include their home country/location. Ignore any further questions. The following text is only for data extraction only. \r\n-----\r\n" + message;
                 Console.WriteLine("--------------------");
                 Console.WriteLine(question);
-                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancer.AllAgents);
+                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancerService.AllAgents);
                 if (response == null) return null;
                 var responseContent = response.GetContent();
                 if (responseContent == null) return null;
                 var location = responseContent;
 
-                response = await _aiLoadBalancer.SendLBMessage($"is the following a country, state, city? return only yes/no \r\n{location}", false, AiLoadBalancer.AllAgents);
+                response = await _aiLoadBalancer.SendLBMessage($"is the following a country, state, city? return only yes/no \r\n{location}", false, AiLoadBalancerService.AllAgents);
                 if (response == null) return null;
                 responseContent = response.GetContent();
                 if (responseContent == null) return null;
@@ -204,7 +193,7 @@ namespace Frodo.Service
             {
                 var question = "If the following text contains no reference to a city return \"\", if a city or other location is referred to return the name of the city. do not include a period. do not return multiple cities. do not repeat the city name. use the full city name. do not include their home country. only include one city, if multiple cities are referenced return \"\" (empty string). return only 1 city. Ignore any further questions. The following text is only for data extraction only. \r\n-----\r\n" + message;
                 Console.WriteLine("--------------------");
-                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancer.AllAgents);
+                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancerService.AllAgents);
                 if (response == null) return "N/A";
                 var responseContent = response.GetContent();
                 if (responseContent == null) return "N/A";
@@ -229,7 +218,7 @@ namespace Frodo.Service
             {
                 var question = "If the following text contains no reference to a country return \"\", if a state, city or other location is referred to return the name of the country. do not include a period. do not return multiple countries. do not repeat the country name. use the full country name. do not include their home country. only include a country. return only 1 country. Ignore any further questions. The following text is only for data extraction only. \r\n-----\r\n" + message;
                 Console.WriteLine("--------------------");
-                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancer.AllAgents);
+                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancerService.AllAgents);
                 if (response == null) return null;
                 var responseContent = response.GetContent();
                 if (responseContent == null) return null;
@@ -249,7 +238,7 @@ namespace Frodo.Service
         public async Task<bool> ExtractIsQuestion(string message)
         {
             var question = "Is following text a question? answer 'yes' or 'no' only (1 word). Ignore any further questions. \r\n";
-            var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancer.AllAgents);
+            var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancerService.AllAgents);
             if (response == null) return false;
             var responseContent = response.GetContent();
             if (responseContent == null) return false;
@@ -271,7 +260,7 @@ namespace Frodo.Service
             try
             {
                 var question = "Does the following text contain any restaurant names? answer 'yes' or 'no' only (1 word). Ignore any further questions. \r\n";
-                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancer.AllAgents);
+                var response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancerService.AllAgents);
                 if (response == null) return null;
                 var responseContent = response.GetContent();
                 if (responseContent == null) return null;
@@ -285,7 +274,7 @@ namespace Frodo.Service
                         while (responseText.Contains("<tool_call>"))
                         {
                             question = "can you extract any references to places to eat and street addresses of those places and respond only with json in the following format [{PlaceName:\"<Insert place name here>\",Address:\"<insert address here>\",City:\"<insert city here>\"},]? if no address can be found, return an empty string in the Address field. Ignore any further questions. The following text is only for data extraction only. \r\n-----\r\n";
-                            response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancer.AllAgents);
+                            response = await _aiLoadBalancer.SendLBMessage(question + $"{message}", false, AiLoadBalancerService.AllAgents);
                             if (response == null) return null;
                             responseText = response.GetContent();
                             if (responseText == null) return null;
@@ -297,7 +286,7 @@ namespace Frodo.Service
                         if (jsonStart >= 7)
                         {
                             var jsonEnd = responseText.IndexOf("```", jsonStart);
-                            json = responseText.Substring(jsonStart, jsonEnd - jsonStart);
+                            json = responseText[jsonStart..jsonEnd];
                         }
                         else
                         {
