+++ Smeagol/Smeagol.sln
@@ -13,6 +13,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gluten.Core", "..\_Library\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gluten.Core.DataProcessing", "..\_Library\Gluten.Core.DataProcessing\Gluten.Core.DataProcessing.csproj", "{953C32A2-D48D-4F76-B429-D03576750D5A}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gluten.Core.LocationProcessing", "..\_Library\Gluten.Core.LocationProcessing\Gluten.Core.LocationProcessing.csproj", "{DABBD827-55AF-48E3-9C87-64145D73CB41}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -35,6 +37,10 @@ Global
 		{953C32A2-D48D-4F76-B429-D03576750D5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{953C32A2-D48D-4F76-B429-D03576750D5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{953C32A2-D48D-4F76-B429-D03576750D5A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DABBD827-55AF-48E3-9C87-64145D73CB41}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DABBD827-55AF-48E3-9C87-64145D73CB41}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DABBD827-55AF-48E3-9C87-64145D73CB41}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DABBD827-55AF-48E3-9C87-64145D73CB41}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -43,6 +49,7 @@ Global
 		{A30A8CEE-AE51-483C-A310-E9CD4FEF9CE6} = {698FCDBB-691D-4BC6-8259-23696249713D}
 		{58D6E65A-CB47-4C65-BDE9-625A66027AE4} = {698FCDBB-691D-4BC6-8259-23696249713D}
 		{953C32A2-D48D-4F76-B429-D03576750D5A} = {698FCDBB-691D-4BC6-8259-23696249713D}
+		{DABBD827-55AF-48E3-9C87-64145D73CB41} = {698FCDBB-691D-4BC6-8259-23696249713D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F1878937-7EA1-4F00-9E15-32D8E47CD067}
