+++ _Library/Gluten.Data/PinCache/TopicPinCache.cs
@@ -5,13 +5,13 @@
         public required string PlaceName { get; set; }
         public required string Label { get; set; }
         public string? Address { get; set; }
-        public string? GeoLatitude { get; set; }
-        public string? GeoLongitude { get; set; }
+        public string GeoLatitude { get; set; } = "";
+        public string GeoLongitude { get; set; } = "";
         public string? MapsUrl { get; set; }
         public string? MetaHtml { get; set; }
         public PinCacheMeta? MetaData { get; set; }
         public bool MetaProcessed { get; set; } = false;
         public string? Country { get; set; }
-        public List<string> SearchStrings { get; set; } = new();
+        public List<string> SearchStrings { get; set; } = [];
     }
 }
